refactor(proto): isolate window serde helpers
Window protobuf helpers obscured the execution plan implementations.
Keeping them feature-gated together makes the shared serialization
invariants easier to maintain without widening their visibility.

Refs #23513
