+++ datafusion/physical-plan/src/windows/mod.rs
@@ -18,6 +18,8 @@
 //! Physical expressions for window functions
 
 mod bounded_window_agg_exec;
+#[cfg(feature = "proto")]
+mod proto;
 mod utils;
 mod window_agg_exec;
